drf_ecs/requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-cors-headers==2.5.3
django-mysql==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.9.2
djangorestframework-simplejwt==4.3.0
djoser==1.7.0
drf-yasg==1.16.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
PyJWT==1.7.1
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
ruamel.yaml==0.16.0
ruamel.yaml.clib==0.1.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3